feat: adding some files gui

+++ BashDBMS/lib/drop_db.sh
@@ -1,10 +1,17 @@
 #!/bin/bash
 
-read -p "Enter Database Name to Drop: " db_name
+db_name=$(zenity --entry \
+    --title="Drop Database" \
+    --text="Enter Database Name to Drop:")
+
+[[ -z "$db_name" ]] && return
 
 if [[ -d "./databases/$db_name" ]]; then
-    rm -r "./databases/$db_name"
-    echo "Database deleted."
+    zenity --question --text="Are you sure you want to delete '$db_name'?"
+    if [[ $? -eq 0 ]]; then
+        rm -r "./databases/$db_name"
+        zenity --info --text="Database deleted."
+    fi
 else
-    echo "Database not found."
+    zenity --error --text="Database not found."
 fi
